Rewritten dhcp to use scapy
Rewritten dhcp to use scapy.
Reason being it could not work when having multiple network interfaces, it would not capture data on 0.0.0.0 from NIC2

It still cannot capture multiple data NICs data but it can work on both NIC1 and NIC2
There may still be some things that needs to be fixed in this commit causing Tk or to not work.